added tests for waver filter 1
