+++ tests/services/case_filters/test_case_filter_wave_1.py
@@ -1,0 +1,312 @@
+import logging
+
+import pytest
+
+from models.blaise.blaise_case_information_model import BlaiseCaseInformationModel
+from models.totalmobile.totalmobile_world_model import TotalmobileWorldModel
+from services.case_filters.case_filter_wave_1 import CaseFilterWave1
+from services.eligible_case_service import EligibleCaseService
+from tests.helpers.get_blaise_case_model_helper import get_populated_case_model
+
+
+@pytest.fixture()
+def service() -> CaseFilterWave1:
+    return CaseFilterWave1()
+
+
+@pytest.fixture()
+def valid_wave_1_case() -> BlaiseCaseInformationModel:
+    return get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="",
+            wave=1,
+            field_case="Y",
+            outcome_code=310,
+            field_region="Region 1",
+        )
+
+
+def test_case_is_eligible_returns_true_only_where_criteria_for_wave_1_is_met(
+    valid_wave_1_case,
+    service: CaseFilterWave1,
+):
+    # act
+    result = service.case_is_eligible(valid_wave_1_case)
+
+    # assert
+    assert result is True
+
+
+def test_case_is_eligible_returns_false_if_the_case_is_not_wave_1(
+    valid_wave_1_case,
+    service: CaseFilterWave1,
+):
+    # arrange
+    case = valid_wave_1_case
+    case.wave = 2
+
+    # act
+    result = service.case_is_eligible(case)
+
+    # assert
+    assert result is False
+
+
+def test_case_is_eligible_returns_false_if_telephone_number_1_has_a_value(
+    valid_wave_1_case,
+    service: CaseFilterWave1,
+):
+    # arrange
+    case = valid_wave_1_case
+    case.contact_details.telephone_number_1 = "3535232"
+
+    # act
+    result = service.case_is_eligible(case)
+
+    # assert
+    assert result is False
+
+
+def test_case_is_eligible_logs_a_message_if_telephone_number_1_has_a_value(
+        valid_wave_1_case,
+        service: CaseFilterWave1,
+        caplog
+):
+    # arrange
+    case = valid_wave_1_case
+    case.contact_details.telephone_number_1 = "3535232"
+
+    # act && assert
+    with caplog.at_level(logging.INFO):
+        service.case_is_eligible(case)
+    assert (
+        "root",
+        logging.INFO,
+        "Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a value set for the field 'telephone_number_1'",
+    ) in caplog.record_tuples
+
+
+
+
+def test_get_eligible_cases_logs_a_message_when_a_case_is_not_eligible_as_telephone_number_2_has_a_value(
+    service: EligibleCaseService,
+    caplog,
+):
+    # arrange
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="2221221",
+            appointment_telephone_number="",
+            wave=1,
+            priority="1",
+            outcome_code=310,
+            field_region="Region 1",
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        "Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a value set for the field 'telephone_number_2'",
+    ) in caplog.record_tuples
+
+
+def test_get_eligible_cases_logs_a_message_when_a_case_is_not_eligible_as_appointment_telephone_number_has_a_value(
+    service: EligibleCaseService,
+    caplog,
+):
+    # arrange
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="2221221",
+            wave=1,
+            priority="1",
+            outcome_code=310,
+            field_region="Region 1",
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        "Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a value set for the field 'appointment_telephone_number'",
+    ) in caplog.record_tuples
+
+
+@pytest.mark.parametrize("test_input", [110, 210, 410])
+def test_get_eligible_cases_logs_a_message_when_a_priority_is_not_in_range(
+    test_input, service: EligibleCaseService, caplog
+):
+    # arrange
+    value_range = [0, 310, 320]
+
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="",
+            wave=1,
+            priority="1",
+            outcome_code=test_input,
+            field_region="Region 1",
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        f"Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a value '{test_input}' outside of the range '{value_range}' set for the field 'outcome_code'",
+    ) in caplog.record_tuples
+
+
+def test_get_eligible_cases_logs_a_message_when_field_case_is_set_to_n(service, caplog):
+    # arrange
+
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="",
+            wave=1,
+            field_case="N",
+            priority="1",
+            outcome_code=310,
+            field_region="Region 1",
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        f"Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a field case value of 'N', not 'Y'",
+    ) in caplog.record_tuples
+
+
+def test_get_eligible_cases_logs_a_message_when_field_case_is_set_to_N_and_priority_is_missing(
+    service: EligibleCaseService,
+    caplog,
+):
+    # arrange
+
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="",
+            wave=1,
+            field_case="N",
+            outcome_code=310,
+            field_region="Region 1",
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        f"Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a field case value of 'N', not 'Y'",
+    ) in caplog.record_tuples
+
+
+def test_get_eligible_cases_logs_a_message_when_field_case_is_set_to_an_empty_string(
+    service: EligibleCaseService,
+    caplog,
+):
+    # arrange
+
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="",
+            wave=1,
+            field_case="",
+            outcome_code=310,
+            field_region="Region 1",
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        f"Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a field case value of '', not 'Y'",
+    ) in caplog.record_tuples
+
+
+@pytest.mark.parametrize("test_input", ["Region 0", "Region 9", "Default"])
+def test_get_eligible_cases_logs_a_message_when_a_field_region_is_not_in_range(
+    test_input, service: EligibleCaseService, caplog
+):
+    # arrange
+    value_range = TotalmobileWorldModel.get_available_regions()
+
+    cases = [
+        get_populated_case_model(
+            case_id="90001",
+            telephone_number_1="",
+            telephone_number_2="",
+            appointment_telephone_number="",
+            wave=1,
+            priority="1",
+            outcome_code=310,
+            field_region=test_input,
+        )
+    ]
+
+    # act
+    result = service.get_eligible_cases(cases)
+
+    # assert
+    assert len(result) == 0
+
+    assert (
+        "root",
+        logging.INFO,
+        f"Case '90001' in questionnaire 'LMS2101_AA1' was not eligible to be sent to Totalmobile as it has a value '{test_input}' outside of the range '{value_range}' set for the field 'field_region'",
+    ) in caplog.record_tuples
